1.专家打分支持小数点后一位 2.系统的价格支持小数点后三位

+++ youyaboot-service/src/main/java/com/magicalcoder/youyaboot/model/CommonSum.java
@@ -29,8 +29,8 @@ public class CommonSum implements Serializable{
     private String pj;
     private String signature;
     private String count;
-    private Integer basePoint;
-    private Integer finalPoint;
+    private double basePoint;
+    private double finalPoint;
     private String date;
 
 }
